requirements.txt
fastapi
uvicorn
strawberry-graphql[fastapi]
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
httpx
python-multipart